+++ src/Products/SiteErrorLog/tests/testSiteErrorLog.py
@@ -18,10 +18,15 @@ import unittest
 
 import transaction
 from Testing.makerequest import makerequest
+import Testing.testbrowser
+import Testing.ZopeTestCase
 from zope.component import adapter, provideHandler
 from zope.event import notify
 from ZPublisher.pubevents import PubFailure
 import Zope2
+import Zope2.App
+
+from ..SiteErrorLog import manage_addErrorLog
 
 from Products.SiteErrorLog.interfaces import IErrorRaisedEvent
 
@@ -177,3 +182,39 @@ class SiteErrorLogTests(unittest.TestCase):
         # Need to make sure that the __error_log__ hook gets cleaned up
         self.app._delObject('error_log')
         self.assertEqual(getattr(self.app, '__error_log__', None), None)
+
+
+class SiteErrorLogUITests(Testing.ZopeTestCase.FunctionalTestCase):
+
+    def setUp(self):
+        super(SiteErrorLogUITests, self).setUp()
+
+        Zope2.App.zcml.load_site(force=True)
+
+        uf = self.app.acl_users
+        uf.userFolderAddUser('manager', 'manager_pass', ['Manager'], [])
+
+        # Why is this neccessary, shouldn't the test get a new app every time?
+        if not hasattr(self.app, 'error_log'):
+            manage_addErrorLog(self.app)
+        self.error_log = self.app.error_log
+
+        self.browser = Testing.testbrowser.Browser()
+        self.browser.login('manager', 'manager_pass')
+        self.browser.open('http://localhost/error_log/manage_main')
+
+    def testSubmitRetainsIgnoredExceptionsUnchanged(self):
+        # Checks the fix for
+        # https://github.com/zopefoundation/Products.SiteErrorLog/issues/13
+        ignoredExceptions = self.browser.getControl(
+            label='Ignored exception types')
+        self.assertEqual(
+            ignoredExceptions.value,
+            'Unauthorized\nNotFound\nRedirect')  # default value
+        ignoredExceptions.value = 'Unauthorized\nFnord'
+        self.assertNotIn('Changed properties', self.browser.contents)
+        self.browser.getControl('Save Changes').click()
+        self.assertIn('Changed properties', self.browser.contents)
+        ignoredExceptions = self.browser.getControl(
+            label='Ignored exception types')
+        self.assertEqual(ignoredExceptions.value, 'Unauthorized\nFnord')